+++ test/bigtest/interactors/error-modal.js
@@ -1,10 +1,16 @@
 import {
-  interactor,
   clickable,
+  interactor,
   isPresent,
+  scoped,
 } from '@bigtest/interactor';
 
-@interactor class ErrorModalInteractor {
+@interactor class ErrorModal {
+  defaultScope = '[data-test-error-modal]';
+
+  content = scoped('[data-test-error-modal-content]');
+  closeButton = scoped('[data-test-error-modal-close-button]');
+
   modalIsPresent = isPresent('[data-test-error-modal]');
   clickCloseBtn = clickable('[data-test-error-modal-close-button]');
 
@@ -13,4 +19,4 @@ import {
   }
 }
 
-export default new ErrorModalInteractor('#OverlayContainer');
+export default ErrorModal;