+++ test/bigtest/interactors/header-dropdown.js
@@ -1,0 +1,11 @@
+import {
+  interactor,
+  clickable,
+} from '@bigtest/interactor';
+
+
+@interactor class HeaderDropdown {
+  click = clickable('[data-test-pane-header-actions-button]');
+}
+
+export default HeaderDropdown;