+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Rename request export names in Action dropdown, make links disabled if no search/filtered results. Refs UIREQ-395.
 * Include effective call number elements in the Request CSV. Refs UIREQ-367.
 * Include effective call number elements in the Hold shelf clearance report. Refs UIREQ-368.
+* Only fetch open loans when loading an item. Fixes UIREQ-353.
 
 ## [1.14.0](https://github.com/folio-org/ui-requests/tree/v1.14.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v1.13.0...v1.14.0)